one last push before turning in

+++ db/schema.sql
@@ -1,7 +1,9 @@
 DROP DATABASE IF EXISTS company_db;
 CREATE DATABASE company_db;
+
 USE company_db;
+
 CREATE TABLE departments(id INT PRIMARY KEY AUTO_INCREMENT, name VARCHAR(30));
 CREATE TABLE managers(id INT PRIMARY KEY AUTO_INCREMENT, first_name VARCHAR (30), last_name VARCHAR (30));
-CREATE TABLE roles(id INT PRIMARY KEY AUTO_INCREMENT, title VARCHAR(30), salary DECIMAL, department_id INT);
+CREATE TABLE roles(id INT PRIMARY KEY AUTO_INCREMENT, title VARCHAR(30), salary DECIMAL(10, 2), department_id INT);
 CREATE TABLE employees(id INT PRIMARY KEY AUTO_INCREMENT, first_name VARCHAR (30), last_name VARCHAR(30), role_id INT, manager_id INT, department_id INT);
